tidy up and update README for go lang
