--- go/.env
@@ -1,10 +1,0 @@
-# THIS IS GO VERSION
-# database
-database_name=your_database_name
-database_password=your_very_secret_password
-database_host=host_to_your_database
-database_port=port_to_your_database
-
-# server
-bearer=Bearer abcdefg
-endpoint=https://www.your_nice_endpoint.com
