--- go/env.example
@@ -1,10 +1,0 @@
-# THIS IS GO VERSION
-# database
-database_name=
-database_password=
-database_host=
-database_port=
-
-# server
-bearer=
-endpoint=
